refactored to have state at the global level, accessible, changeable from different pages
